Monalisa

+++ Backend/src/main/java/com/bloghorizon/backend/dtos/UserResponse.java
@@ -1,17 +1,17 @@
 package com.bloghorizon.backend.dtos;
 
 public class UserResponse {
-    private final int id;
+    private final Long id;
     private final String username;
     private final String email;
 
-    public UserResponse(int id, String username, String email) {
+    public UserResponse(Long id, String username, String email) {
         this.id = id;
         this.username = username;
         this.email = email;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
